tools/synthetic-adsb/requirements.txt
flask==3.0.2
requests==2.31.0
flask-cors==4.0.0 
python-dotenv==1.0.1